# Refactor character handling in CharExtensions and ReadOnlyMemoryExtensions