SQL parser optimized: achieved 15-25% performance gains and enhanced stability across all query lengths.

- Changed data structures from HashSet<char> to char[] for improved efficiency.
- Updated methods to use loops instead of Contains for better performance.

## before
| Method                         | Mean       | Error     | StdDev    |
|------------------------------- |-----------:|----------:|----------:|
| Carbunqlex_Parse_Short         |   7.530 μs | 0.0809 μs | 0.0717 μs |
| Carbunqlex_ParseOnly_Short     |   6.770 μs | 0.0197 μs | 0.0154 μs |
| Carbunqlex_Parse_Middle        |  21.180 μs | 0.0782 μs | 0.0610 μs |
| Carbunqlex_ParseOnly_Middle    |  19.101 μs | 0.1167 μs | 0.1034 μs |
| Carbunqlex_Parse_Long          |  40.113 μs | 0.2124 μs | 0.1774 μs |
| Carbunqlex_ParseOnly_Long      |  38.414 μs | 0.1544 μs | 0.1369 μs |
| Carbunqlex_Parse_SuperLong     |  85.758 μs | 0.8017 μs | 0.7107 μs |
| Carbunqlex_ParseOnly_SuperLong |  62.131 μs | 0.7448 μs | 0.6966 μs |

## after
| Method                         | Mean      | Error     | StdDev    |
|------------------------------- |----------:|----------:|----------:|
| Carbunqlex_Parse_Short         |  7.578 μs | 0.0345 μs | 0.0306 μs |
| Carbunqlex_ParseOnly_Short     |  6.487 μs | 0.0194 μs | 0.0162 μs |
| Carbunqlex_Parse_Middle        | 20.482 μs | 0.1782 μs | 0.1667 μs |
| Carbunqlex_ParseOnly_Middle    | 18.839 μs | 0.0808 μs | 0.0716 μs |
| Carbunqlex_Parse_Long          | 38.988 μs | 0.3077 μs | 0.2878 μs |
| Carbunqlex_ParseOnly_Long      | 35.014 μs | 0.3216 μs | 0.2851 μs |
| Carbunqlex_Parse_SuperLong     | 82.868 μs | 0.5076 μs | 0.4748 μs |
| Carbunqlex_ParseOnly_SuperLong | 55.696 μs | 0.4335 μs | 0.4055 μs |

## after2
| Method                         | Mean      | Error     | StdDev    |
|------------------------------- |----------:|----------:|----------:|
| Carbunqlex_Parse_Short         |  6.199 μs | 0.0160 μs | 0.0134 μs |
| Carbunqlex_ParseOnly_Short     |  5.432 μs | 0.0296 μs | 0.0247 μs |
| Carbunqlex_Parse_Middle        | 17.340 μs | 0.0881 μs | 0.0824 μs |
| Carbunqlex_ParseOnly_Middle    | 15.854 μs | 0.1539 μs | 0.1440 μs |
| Carbunqlex_Parse_Long          | 33.004 μs | 0.1898 μs | 0.1683 μs |
| Carbunqlex_ParseOnly_Long      | 29.729 μs | 0.2178 μs | 0.1930 μs |
| Carbunqlex_Parse_SuperLong     | 72.868 μs | 0.7630 μs | 0.6371 μs |
| Carbunqlex_ParseOnly_SuperLong | 46.911 μs | 0.2853 μs | 0.2669 μs |